✨feat: game over and victory screen

+++ src/model/Ball.hpp
@@ -3,20 +3,20 @@
 
 struct BallPositions
 {
-    float x = SCREEN_WIDTH/2;
-    float y = BALL_BOTTOM_POSITION;
+    float x = static_cast<float>(SCREEN_WIDTH / 2);
+    float y = static_cast<float>(BALL_BOTTOM_POSITION);
 };
 
 struct BallDirection
 {
     float x = 0;
-    float y = -BALL_DEFAULT_SPEED;
+    float y = static_cast<float>(-BALL_DEFAULT_SPEED);
 };
 
 struct BallParamaters
 {
-    float radius = BALL_DEFAULT_RADIUS;
-    float speed = BALL_DEFAULT_SPEED;
+    float radius = static_cast<float>(BALL_DEFAULT_RADIUS);
+    float speed = static_cast<float>(BALL_DEFAULT_SPEED);
 };
 
 class Ball
@@ -26,7 +26,7 @@ public:
     void update();
     bool isLost() const;
     void reset();
-    void setPosition(float x, float y = BALL_BOTTOM_POSITION);
+    void setPosition(float x, float y = static_cast<float>(BALL_BOTTOM_POSITION));
     void setDirection(float dx, float dy);
     void setRadius(float radius);
     void setBallSpeed(float speed);
