+++ src/model/Racket.hpp
@@ -3,15 +3,15 @@
 
 struct RacketPositions
 {
-    float x = SCREEN_WIDTH/2 - (RACKET_DEFAULT_WIDTH / 2);
-    float y = RACKET_BOTTOM_POSITION;
+    float x = static_cast<float>(SCREEN_WIDTH / 2 - (RACKET_DEFAULT_WIDTH / 2));
+    float y = static_cast<float>(RACKET_BOTTOM_POSITION);
 };
 
 struct RacketParameters
 {
-    int width = RACKET_DEFAULT_WIDTH;
-    int height = RACKET_DEFAULT_HEIGHT;
-    float speed = RACKET_DEFAULT_SPEED;
+    int width = (RACKET_DEFAULT_WIDTH);
+    int height = (RACKET_DEFAULT_HEIGHT);
+    int speed = (RACKET_DEFAULT_SPEED);
 };
 
 class Racket
@@ -24,7 +24,7 @@ public:
     void setPosition(float x);
     const RacketPositions &getPositions() const;
     const RacketParameters &getParameters() const;
-    void setParameters(int width, int height, float speed); // if in the future we want to customize the racket
+    void setParameters(int width, int height, int speed); // if in the future we want to customize the racket
 
 private:
     RacketPositions positions_;
